Migrated models to DBFlow
- Goal value is now a double
- The request code and extra stored for the PendingIntent is the goal ID
- When a notification arrives, the goal is obtained from the db
- TODO: Make DBFlow work with Robolectric

+++ app/src/main/java/com/nispok/fitcoach/dialogs/GoalDialogFragment.java
@@ -98,7 +98,7 @@ public class GoalDialogFragment extends DialogFragment {
         okButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                goal.setValue(Integer.parseInt(goalInput.getText().toString()));
+                goal.setValue(Double.parseDouble(goalInput.getText().toString()));
                 listener.onGoalChanged(goal);
                 GoalDialogFragment.this.dismiss();
             }
